+++ src/main/java/com/springboot/minimartapi/product/ProductEditionDto.java
@@ -1,7 +1,4 @@
 package com.springboot.minimartapi.product;
-
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
 import lombok.Builder;
 
 @Builder
@@ -12,7 +9,7 @@ public record ProductEditionDto(
 
         String productDescription,
 
-        Float price,
+        Double price,
 
         Category category
